+++ test2/data/conversion-delta-temperature.ts
@@ -1,6 +1,6 @@
 import { ConversionTest } from "./conversion-test";
-import * as Units from "../src/units";
-import { DeltaTemperature } from "../src/quantity";
+import * as Units from "../../src/units";
+import { DeltaTemperature } from "../../src/quantity";
 
 export const tests: ReadonlyArray<ConversionTest<DeltaTemperature>> = [
   {
