+++ test2/data/conversion-energy.ts
@@ -1,6 +1,6 @@
 import { ConversionTest } from "./conversion-test";
-import * as Units from "../src/units";
-import { Energy } from "../src/quantity";
+import * as Units from "../../src/units";
+import { Energy } from "../../src/quantity";
 
 export const tests: ReadonlyArray<ConversionTest<Energy>> = [
   {
