+++ test2/data/conversion-length.ts
@@ -1,6 +1,6 @@
 import { ConversionTest } from "./conversion-test";
-import * as Units from "../src/units";
-import { Length } from "../src/quantity";
+import * as Units from "../../src/units";
+import { Length } from "../../src/quantity";
 
 export const tests: ReadonlyArray<ConversionTest<Length>> = [
   {
